requirements.txt
pandas~=2.3.2
numpy~=2.2.6
matplotlib~=3.10.6
umap-learn~=0.5.9.post2
scikit-learn~=1.7.2
pytest~=8.4.2